requirements.txt
# 필수 패키지
langgraph>=0.0.20
google-generativeai>=0.3.0
pydantic>=2.0.0

# 선택사항
aiofiles>=23.0.0  # 비동기 파일 처리
python-dotenv>=1.0.0  # 환경 변수 관리
